flask_api/requirements.txt
# requirements.txt
Flask==2.1.1
torch==1.10.1
torchvision==0.11.2
transformers==4.17.0
Pillow==9.0.1
scikit-learn==1.0.2
